+++ Source/Android/src/org/dolphinemu/dolphinemu/about/AboutActivity.java
@@ -1,3 +1,9 @@
+/**
+ * Copyright 2014 Dolphin Emulator Project
+ * Licensed under GPLv2
+ * Refer to the license.txt file included.
+ */
+
 package org.dolphinemu.dolphinemu.about;
 
 import org.dolphinemu.dolphinemu.R;
@@ -10,17 +16,10 @@ import android.app.Activity;
 import android.app.Fragment;
 import android.app.FragmentManager;
 import android.app.FragmentTransaction;
-import android.content.Context;
 import android.os.Bundle;
 import android.support.v13.app.FragmentPagerAdapter;
 import android.support.v4.view.ViewPager;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ArrayAdapter;
-import android.widget.TextView;
 
-import java.util.List;
 
 /**
  * Activity for the about menu, which displays info
@@ -30,75 +29,6 @@ public final class AboutActivity extends Activity implements TabListener
 {
 	private ViewPager viewPager;
 	private final EGLHelper eglHelper = new EGLHelper(EGLHelper.EGL_OPENGL_ES2_BIT);
-
-	// Represents an item in the multiple About fragments.
-	public static final class AboutFragmentItem
-	{
-		private final String title;
-		private final String subtitle;
-
-		public AboutFragmentItem(String title, String subtitle)
-		{
-			this.title = title;
-			this.subtitle = subtitle;
-		}
-
-		public String getTitle()
-		{
-			return title;
-		}
-
-		public String getSubTitle()
-		{
-			return subtitle;
-		}
-	}
-
-	// The adapter that manages the displaying of items in multiple About fragments.
-	public static final class InfoFragmentAdapter extends ArrayAdapter<AboutFragmentItem>
-	{
-		private final int id;
-		private final List<AboutFragmentItem> items;
-
-		public InfoFragmentAdapter(Context ctx, int id, List<AboutFragmentItem> items)
-		{
-			super(ctx, id, items);
-
-			this.id = id;
-			this.items = items;
-		}
-
-		@Override
-		public AboutFragmentItem getItem(int index)
-		{
-			return items.get(index);
-		}
-
-		@Override
-		public View getView(int position, View convertView, ViewGroup parent)
-		{
-			if (convertView == null)
-			{
-				LayoutInflater vi = LayoutInflater.from(getContext());
-				convertView = vi.inflate(id, parent, false);
-			}
-
-			final AboutFragmentItem item = items.get(position);
-			if (item != null)
-			{
-				TextView title    = (TextView) convertView.findViewById(R.id.AboutItemTitle);
-				TextView subtitle = (TextView) convertView.findViewById(R.id.AboutItemSubTitle);
-
-				if (title != null)
-					title.setText(item.getTitle());
-
-				if (subtitle != null)
-					subtitle.setText(item.getSubTitle());
-			}
-
-			return convertView;
-		}
-	}
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState)
